requirements.txt
torch
transformers
sentence-transformers
pandas
scikit-learn
faiss-cpu